ext4: correctly handle queries for metadata mappings
commit 46c22a8bb4cb03211da1100d7ee4a2005bf77c70 upstream.

Currently, our handling of metadata is _ambiguous_ in some scenarios,
that is, we end up returning unknown if the range only covers the
mapping partially.

For example, in the following case:

$ xfs_io -c fsmap -d

 0: 254:16 [0..7]: static fs metadata 8
 1: 254:16 [8..15]: special 102:1 8
 2: 254:16 [16..5127]: special 102:2 5112
 3: 254:16 [5128..5255]: special 102:3 128
 4: 254:16 [5256..5383]: special 102:4 128
 5: 254:16 [5384..70919]: inodes 65536
 6: 254:16 [70920..70967]: unknown 48
 ...

$ xfs_io -c fsmap -d 24 33

 0: 254:16 [24..39]: unknown 16 <--- incomplete reporting

$ xfs_io -c fsmap -d 24 33 (With patch)

  0: 254:16 [16..5127]: special 102:2 5112

This is because earlier in ext4_getfsmap_meta_helper, we end up ignoring
any extent that starts before our queried range, but overlaps it. While
the man page [1] is a bit ambiguous on this, this fix makes the output
make more sense since we are anyways returning an "unknown" extent. This
is also consistent to how XFS does it:

$ xfs_io -c fsmap -d

 ...
 6: 254:16 [104..127]: free space 24
 7: 254:16 [128..191]: inodes 64
 ...

$ xfs_io -c fsmap -d 137 150

 0: 254:16 [128..191]: inodes 64  <-- full extent returned

 [1] https://man7.org/linux/man-pages/man2/ioctl_getfsmap.2.html

Reported-by: Ritesh Harjani (IBM) <ritesh.list@gmail.com>
Cc: stable@kernel.org
Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>
Message-ID: <023f37e35ee280cd9baac0296cbadcbe10995cab.1757058211.git.ojaswin@linux.ibm.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Ulrich Hecht <uli@kernel.org>
Signed-off-by: Youffx <asa.jazal@gmail.com>